+++ src/pageEditor/EditorLayout.tsx
@@ -16,7 +16,7 @@
  */
 
 import React from "react";
-import ModListingPanel from "@/pageEditor/sidebar/ModListingPanel";
+import ModListingPanel from "@/pageEditor/modListingPanel/ModListingPanel";
 import { useSelector } from "react-redux";
 import useFlags from "@/hooks/useFlags";
 import Modals from "./Modals";
