+++ src/pageEditor/Modals.tsx
@@ -17,10 +17,10 @@
 
 import AddBrickModal from "@/pageEditor/modals/addBrickModal/AddBrickModal";
 import React from "react";
-import AddToModModal from "./sidebar/modals/AddToModModal";
-import CreateModModal from "./sidebar/modals/CreateModModal";
-import MoveFromModModal from "./sidebar/modals/MoveFromModModal";
-import SaveAsNewModModal from "./sidebar/modals/SaveAsNewModModal";
+import AddToModModal from "@/pageEditor/modListingPanel/modals/AddToModModal";
+import CreateModModal from "@/pageEditor/modListingPanel/modals/CreateModModal";
+import MoveFromModModal from "@/pageEditor/modListingPanel/modals/MoveFromModModal";
+import SaveAsNewModModal from "@/pageEditor/modListingPanel/modals/SaveAsNewModModal";
 import SaveDataIntegrityErrorModal from "@/pageEditor/panes/save/SaveDataIntegrityErrorModal";
 
 const Modals: React.FunctionComponent = () => (
