+++ src/pageEditor/modListingPanel/ActionMenu.tsx
@@ -16,7 +16,7 @@
  */
 
 import React from "react";
-import SaveButton from "@/pageEditor/sidebar/SaveButton";
+import SaveButton from "@/pageEditor/modListingPanel/SaveButton";
 import {
   faClone,
   faFileExport,
