+++ src/pageEditor/modListingPanel/ActivatedModComponentListItem.test.tsx
@@ -18,7 +18,7 @@
 import React from "react";
 import { render } from "@/pageEditor/testHelpers";
 import { actions as editorActions } from "@/pageEditor/slices/editorSlice";
-import ActivatedModComponentListItem from "@/pageEditor/sidebar/ActivatedModComponentListItem";
+import ActivatedModComponentListItem from "@/pageEditor/modListingPanel/ActivatedModComponentListItem";
 import { modComponentFactory } from "@/testUtils/factories/modComponentFactories";
 import { formStateFactory } from "@/testUtils/factories/pageEditorFactories";
 import { screen, waitFor } from "@testing-library/react";
