+++ src/tsconfig.strictNullChecks.json
@@ -1115,25 +1115,25 @@
     "./pageEditor/panes/save/saveHelpers.ts",
     "./pageEditor/panes/save/savingModComponentSlice.ts",
     "./pageEditor/protocol.ts",
-    "./pageEditor/sidebar/ActionMenu.stories.tsx",
-    "./pageEditor/sidebar/ActionMenu.tsx",
-    "./pageEditor/sidebar/ActivatedModComponentListItem.tsx",
-    "./pageEditor/sidebar/AddStarterBrickButton.tsx",
-    "./pageEditor/sidebar/ModComponentIcons.test.tsx",
-    "./pageEditor/sidebar/ModComponentIcons.tsx",
-    "./pageEditor/sidebar/HomeButton.tsx",
-    "./pageEditor/sidebar/ModListItem.tsx",
-    "./pageEditor/sidebar/ReloadButton.tsx",
-    "./pageEditor/sidebar/SaveButton.stories.tsx",
-    "./pageEditor/sidebar/SaveButton.tsx",
-    "./pageEditor/sidebar/arrangeSidebarItems.test.ts",
-    "./pageEditor/sidebar/arrangeSidebarItems.ts",
-    "./pageEditor/sidebar/common.ts",
-    "./pageEditor/sidebar/filterSidebarItems.test.ts",
-    "./pageEditor/sidebar/filterSidebarItems.ts",
-    "./pageEditor/sidebar/modals/AddToModModal.tsx",
-    "./pageEditor/sidebar/modals/MoveFromModModal.tsx",
-    "./pageEditor/sidebar/modals/SaveAsNewModModal.tsx",
+    "./pageEditor/modListingPanel/ActionMenu.stories.tsx",
+    "./pageEditor/modListingPanel/ActionMenu.tsx",
+    "./pageEditor/modListingPanel/ActivatedModComponentListItem.tsx",
+    "./pageEditor/modListingPanel/AddStarterBrickButton.tsx",
+    "./pageEditor/modListingPanel/ModComponentIcons.test.tsx",
+    "./pageEditor/modListingPanel/ModComponentIcons.tsx",
+    "./pageEditor/modListingPanel/HomeButton.tsx",
+    "./pageEditor/modListingPanel/ModListItem.tsx",
+    "./pageEditor/modListingPanel/ReloadButton.tsx",
+    "./pageEditor/modListingPanel/SaveButton.stories.tsx",
+    "./pageEditor/modListingPanel/SaveButton.tsx",
+    "./pageEditor/modListingPanel/arrangeSidebarItems.test.ts",
+    "./pageEditor/modListingPanel/arrangeSidebarItems.ts",
+    "./pageEditor/modListingPanel/common.ts",
+    "./pageEditor/modListingPanel/filterSidebarItems.test.ts",
+    "./pageEditor/modListingPanel/filterSidebarItems.ts",
+    "./pageEditor/modListingPanel/modals/AddToModModal.tsx",
+    "./pageEditor/modListingPanel/modals/MoveFromModModal.tsx",
+    "./pageEditor/modListingPanel/modals/SaveAsNewModModal.tsx",
     "./pageEditor/slices/editorSelectors.ts",
     "./pageEditor/slices/editorSlice.ts",
     "./pageEditor/slices/editorSliceHelpers.ts",
